Replace SpaceWeatherPhenomenon with EnumSpaceWeatherPhenomenon renaming it as SpaceWeatherPhenomenon.
