+++ src/main/java/fi/fmi/avi/model/swx/immutable/SpaceWeatherAdvisoryImpl.java
@@ -12,7 +12,6 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import fi.fmi.avi.model.swx.AdvisoryNumber;
-import fi.fmi.avi.model.swx.EnumSpaceWeatherPhenomenon;
 import fi.fmi.avi.model.swx.IssuingCenter;
 import fi.fmi.avi.model.swx.NextAdvisory;
 import fi.fmi.avi.model.swx.SpaceWeatherAdvisory;
@@ -96,13 +95,12 @@ public abstract class SpaceWeatherAdvisoryImpl implements SpaceWeatherAdvisory, 
                         .setNextAdvisory(NextAdvisoryImpl.immutableCopyOf(value.getNextAdvisory()));
 
                 retval.addAllPhenomena(value.getPhenomena().stream()//
-                        .map(p -> EnumSpaceWeatherPhenomenon.from(p.getType(), p.getSeverity())));
+                        .map(p -> SpaceWeatherPhenomenon.from(p.getType(), p.getSeverity())));
                 retval.addAllAnalyses(value.getAnalyses().stream().map(SpaceWeatherAdvisoryAnalysisImpl::immutableCopyOf));
                 return retval;
             }
         }
 
-        @JsonDeserialize(contentAs = EnumSpaceWeatherPhenomenon.class)
         public Builder addAllPhenomena(final List<SpaceWeatherPhenomenon> elements) {
             return super.addAllPhenomena(elements);
         }
